Enhance docker usage
1)Moved configs from run_docker.sh to common.config
2)Logic to mount the model directory built into run_docker.sh
3)Updated README
4)Included logs for docker image build failure and pull failure

+++ run_docker.sh
@@ -4,50 +4,69 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-DOCKER_REGISTRY="registry.gitlab.arm.com/arm-reference-solutions/build-scripts"
-DOCKER_IMAGE="tc_image"
-TAG=20.04 #Ubuntu version
 SCRIPT_DIR="$(realpath --no-symlinks "$(dirname "${BASH_SOURCE[0]}")")"
+source $SCRIPT_DIR/config/common.config
 WORK_DIR="$(dirname "${SCRIPT_DIR}")" #path to the directory into which TC stack is cloned
 BUILD_CMD="build_image"
 PULL_CMD="pull_image"
 RUN_MODEL="run_model"
 BLUE="\e[94m"
+YELLOW="\e[33m"
+RED="\e[31m"
 NC="\e[0m"
 
 verify_img()
 {
-    img_digest_reg="c1d86091676488183c0fe3de16e69395e5a5dee2d214d4beafb04228ee8604ed"
+    img_digest_reg="a058002ef161c4733ef1adac8aed46486918f3c9ff938395e04795fe3a939b35"
     img_digest_local=$(docker images --format "{{.Repository}}:{{.Tag}}:{{.Digest}}" |grep $DOCKER_IMAGE |cut -d":" -f4)
     [ $img_digest_reg == $img_digest_local ] && exit 0
-    echo "Image verification failed!!" && exit 1
+    echo -e "${YELLOW}Image verification failed!! ${NC}" && exit 1
+}
+
+pull_img()
+{
+    docker pull $DOCKER_REGISTRY/$DOCKER_IMAGE:$TAG && return 0
+    echo -e "${YELLOW}Cannot pull Docker image from Container registry!! ${NC}"
+    exit 1
+}
+build_img()
+{
+  docker build --no-cache --build-arg version=$TAG -t $DOCKER_REGISTRY/$DOCKER_IMAGE:$TAG docker/ && return 0
+  echo -e "${RED}ERROR:Docker local build failed ${NC}"
+  exit 1
 }
 
 if [[ "$1" == "$PULL_CMD" ]]; then
 
     echo -e "${BLUE}INFO: PULLING DOCKER IMAGE FROM REGISTRY ${NC}"
-    docker pull $DOCKER_REGISTRY/$DOCKER_IMAGE:$TAG
+    pull_img ;
     verify_img ;
 
 elif [[ "$1" == "$BUILD_CMD" ]]; then
 
     #To build docker image locally
     echo -e "${BLUE}INFO: BUILDING DOCKER IMAGE LOCALLY ${NC}"
-    docker build --no-cache --build-arg version=$TAG -t $DOCKER_REGISTRY/$DOCKER_IMAGE:$TAG docker/
+    build_img ;
 
 elif [[ "$1" == "$RUN_MODEL" ]]; then
 
-   shift;
-   echo -e "${BLUE}INFO: RUNNING FVP IN DOCKER CONTAINER ${NC}"
+    shift;
+    if [ -n "$MODEL_PATH" ] ; then
 
-   lic_opts=" -e LM_LICENSE_FILE=$LM_LICENSE_FILE \
+        echo -e "${BLUE}INFO: RUNNING FVP IN DOCKER CONTAINER ${NC}"
+
+        lic_opts=" -e LM_LICENSE_FILE=$LM_LICENSE_FILE \
               -e ARMLMD_LICENSE_FILE=$ARMLMD_LICENSE_FILE"
 
-   docker run --rm --net=host --mount type=bind,source=$WORK_DIR,target=$WORK_DIR \
-   $lic_opts \
-   --workdir /$WORK_DIR/run-scripts/tc2/ \
-   --ipc=host --env="DISPLAY" -e TERM=$TERM \
-   --user $(id -u):$(id -g) --volume="$HOME/.Xauthority:/home/.Xauthority:rw" -v /tmp/.X11-unix:/tmp/.X11-unix -it $DOCKER_REGISTRY/$DOCKER_IMAGE:$TAG $WORK_DIR/run-scripts/tc2/run_model.sh $@
+        docker run --rm --net=host --mount type=bind,source=$WORK_DIR,target=$WORK_DIR \
+            $lic_opts \
+            --workdir /$WORK_DIR/run-scripts/tc2/ \
+            -v $MODEL_PATH:$MODEL_PATH \
+            --ipc=host --env="DISPLAY" -e TERM=$TERM \
+            --user $(id -u):$(id -g) --volume="$HOME/.Xauthority:/home/.Xauthority:rw" -v /tmp/.X11-unix:/tmp/.X11-unix -it $DOCKER_REGISTRY/$DOCKER_IMAGE:$TAG $WORK_DIR/run-scripts/tc2/run_model.sh $@
+    else
+        echo -e "${RED}Please provide model parent directory path ${NC}"
+    fi
 
 else
 
@@ -64,9 +83,9 @@ else
     echo -e "${BLUE}INFO: ENTERING DOCKER CONTAINER ${NC}"
 
     docker run --rm --net=host --mount type=bind,source=$WORK_DIR,target=$WORK_DIR \
-    $env_opts \
-    --workdir /$SCRIPT_DIR \
-    --user $(id -u):$(id -g) -it $DOCKER_REGISTRY/$DOCKER_IMAGE:$TAG $PWD/$@
+        $env_opts \
+        --workdir /$SCRIPT_DIR \
+        --user $(id -u):$(id -g) -it $DOCKER_REGISTRY/$DOCKER_IMAGE:$TAG $PWD/$@
 
     echo -e "${BLUE}INFO: EXITING DOCKER CONTAINER ${NC}"
 fi
